# openapi.yaml
openapi: 3.0.3
info:
  title: vulnShop API
  description: |
    Deliberately vulnerable API for DAST validation.

    WARNING: This API intentionally contains authentication, authorization,
    injection, information disclosure, and JWT vulnerabilities.
  version: 1.0.0
servers:
  - url: http://localhost:5000
tags:
  - name: Authentication
  - name: Users
  - name: Orders
  - name: Products
  - name: Administration
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    LoginRequest:
      type: object
      required: [username, password]
      properties:
        username: {type: string, example: alice}
        password: {type: string, example: password1}
    LoginResponse:
      type: object
      properties:
        token: {type: string}
        token_type: {type: string, example: Bearer}
        user: {type: object}
    UserUpdate:
      type: object
      properties:
        username: {type: string}
        password: {type: string}
        role: {type: string, example: admin}
        email: {type: string}
        card_number: {type: string}
        api_key: {type: string}
paths:
  /health:
    get:
      tags: [Authentication]
      summary: Health check
      responses: {"200": {description: Application is healthy}}
  /api/login:
    post:
      tags: [Authentication]
      summary: API login
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/LoginRequest'}
      responses:
        "200": {description: Successful login, content: {application/json: {schema: {$ref: '#/components/schemas/LoginResponse'}}}}
        "401": {description: Invalid credentials}
        "500": {description: Database error}
  /api/me:
    get:
      tags: [Users]
      summary: Return JWT claims
      security: [{bearerAuth: []}]
      responses:
        "200": {description: JWT claims}
        "401": {description: Authentication failure}
  /api/users/{user_id}:
    get:
      tags: [Users]
      summary: Get user
      security: [{bearerAuth: []}]
      parameters:
        - {name: user_id, in: path, required: true, schema: {type: integer}, example: 2}
      responses:
        "200": {description: User data}
        "401": {description: Authentication failure}
        "404": {description: User not found}
    put:
      tags: [Users]
      summary: Update user
      security: [{bearerAuth: []}]
      parameters:
        - {name: user_id, in: path, required: true, schema: {type: integer}, example: 2}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/UserUpdate'}
            example: {email: attacker@example.com, role: admin}
      responses:
        "200": {description: User updated}
        "400": {description: Invalid request}
        "401": {description: Authentication failure}
        "500": {description: Verbose database error}
    patch:
      tags: [Users]
      summary: Partially update user
      security: [{bearerAuth: []}]
      parameters:
        - {name: user_id, in: path, required: true, schema: {type: integer}}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/UserUpdate'}
      responses:
        "200": {description: User updated}
        "400": {description: Invalid request}
        "401": {description: Authentication failure}
  /api/orders/{order_id}:
    get:
      tags: [Orders]
      summary: Get order
      security: [{bearerAuth: []}]
      parameters:
        - {name: order_id, in: path, required: true, schema: {type: integer}, example: 1}
      responses:
        "200": {description: Order data}
        "401": {description: Authentication failure}
        "404": {description: Order not found}
  /api/products:
    get:
      tags: [Products]
      summary: Search products
      security: [{bearerAuth: []}]
      parameters:
        - {name: search, in: query, required: false, schema: {type: string}, example: scanner}
      responses:
        "200": {description: Product results}
        "401": {description: Authentication failure}
        "500": {description: Verbose SQL error}
  /api/admin/users:
    get:
      tags: [Administration]
      summary: Administrative user listing
      security: [{bearerAuth: []}]
      responses:
        "200": {description: User list}
        "401": {description: Authentication failure}
        "403": {description: Admin role required}